requirements.txt
pandas
polars
numpy
black
torch
ipykernel
ipywidgets
pathlib
seaborn
dill
optuna
deap
pandas-datareader
scikit-learn
mlflow
pyyaml
tiktoken
transformers
faiss-cpu
datasets
retro-pytorch
sacremoses
newspaper3k
news-please
langchain
clean-text
gensim
fasttext
sentence-transformers
python-dotenv
spacy
#bertopic # downgrade numpy give crash other script!
#numpy==1.21.4 # this downgrade was do to need of bertopic-numba
#numba==0.53
dash
pymongo
newsapi-python
glob2